BREAKING : automate and centralize feature and parameter declaration
Because it was getting very cumbersome and error-prone to declare new 
features, this seems to me a necessary step.

The introduction of an automatic naming scheme for options has the 
following consequences : 

- it is now possible to disable strobing (--no-strobe)
- --strobe, which is on by default, is equivalent to "-" — which I'm 
pretty sure is horrible practice :)
- "-rgb" is no longer a valid alias for "--rgb-shift", being replaced by 
"-rs"
- the RGB shift filter can receive "pause" and "active" options, and the 
zoom filter can receive an "every" option ; however, absent all 
underlying implementation of those in the corresponding functions, they 
are without effect for the moment.

Next, output name generation should also be automated in a similar 
fashion, though that may prove more difficult.