Refactor : make Feature callable, simplifying feature addition
Now, the main script only has to call on every feature as provided by
the preexisting feature list. This means that a feature only needs to be
declared in said list, then implemented as a filter, and it
automatically becomes available.

The one shady part of this is an inexplicable KeyError in strobify.py
when attempting to read the locals() dict for keys that are absolutely
in it, resulting in an ugly hack. Besides this, everything works just as
intended.