Automate feature setting registration through the FeatureSetting type
This is not a very pretty refactor.

FeatureDefaultSettings is gone. Now, a glorified dict type called 
FeatureSettingDefaultValues takes its place, with a check on the 
provided keys to see if they all match declared settings.

Other than this, the main point of frustration is the way setting 
defaults are handled. Previously, they were inlined in the parser 
argument declaration ; now, the FeatureSetting type has a "default" 
field, just like FeatureParameter...which makes for two levels of 
default values : one here, and one in Feature through 
FeatureSettingDefaultValues.

The way this is handled is by having a utility method on Feature that 
returns None if no Feature-level default is found for a given setting ; 
the default then falls back on that provided by FeatureSetting. This is 
convoluted, and I don't like it.