Add afterimages feature (still buggy)
This was a tricky filter because its alpha has to be accounted for *in* 
the filterexpr. It works by tpadding overlaid clones, with each clone 
being applied using the provided alpha.

This required the introduction of the FeatureCombineMode enum, used to 
replace the simple string system used up until this point. Along with 
the preexisting "merge" and "overlay" options, this enum introduces the 
"pre-merged" mode, which always skips applying alpha post-filterexpr no 
matter its value, and instead requires the feature filter to take alpha 
as a parameter to deal with it.

The filter works as expected, except that it necessarily stalls the 
video on activation while the overlays are waiting to start moving. This 
is because tpad only has "clone" (which stalls with the first frame) or 
"add" (which adds a solid RGB color, without alpha).